backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
requests==2.26.0
beautifulsoup4==4.10.0
apscheduler==3.8.1
python-dotenv==0.19.1
gunicorn==20.1.0
schedule==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
flask-bcrypt==1.0.1
PyJWT==2.3.0
Flask-Mail==0.9.1
itsdangerous==2.0.1
Flask-Limiter==2.6.0
SQLAlchemy==1.4.27
psycopg2-binary==2.9.3
alembic==1.7.5 